added options ok status for all api routes

+++ pages/api/daily/delete.ts
@@ -6,7 +6,14 @@ connectDb()
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 	try {
+
+		if (req.method == "OPTIONS") {
+			res.status(200).end()
+			return
+		}
+
 		const { id } = req.body
+		
 
 		const task = await dailyModel.deleteOne({ _id: id })
