+++ pages/api/task/delete.ts
@@ -6,6 +6,11 @@ connectDb()
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 	try {
+		if (req.method == "OPTIONS") {
+			res.status(200).end()
+			return
+		}
+
 		const { id } = req.body
 
 		const task = await taskModel.deleteOne({ _id: id })
